nest_louds_trie: change `build_link()` param type
Param `nextLinkVec` of func `build_link` should be
`valvec<size_t>`, historically, it was `valvec<index_t>`.

Now we store high freq frag into core, this may cause offset
and len can't fit into `index_t` when index_t is uint32

+++ src/terark/zbs/nest_louds_trie_blob_store.cpp
@@ -43,7 +43,7 @@ template<class NestLoudsTrie>
 void
 NestLoudsTrieBlobStore<NestLoudsTrie>::
 build_from(SortableStrVec& strVec, const NestLoudsTrieConfig& conf) {
-	valvec<uint32_t> idvec;
+	valvec<size_t> idvec;
 	AbstractBlobStore::m_unzipSize = strVec.str_size();
 	if (conf.flags[conf.optUseDawgStrPool]) {
 		NestLoudsTrie::build_strpool2(strVec, idvec, conf);
